使用自动生成的课程接口，添加遗漏的接口设计

+++ src/main/java/cn/edu/seu/herald/ws/api/impl/CurriculumServiceImpl.java
@@ -112,7 +112,7 @@ class CurriculumServiceImpl extends AbstractXmlService
         Collections.sort(toBeOrdered, new AttendanceComparator());
         for (Attendance attendance : toBeOrdered) {
             Period period = attendance.getPeriod();
-            int from = period.getFrom();
+            int from = period.getFrom().intValue();
             if (from > startFrom) {
                 return attendance;
             }
@@ -124,7 +124,8 @@ class CurriculumServiceImpl extends AbstractXmlService
             implements Comparator<Attendance> {
 
         public int compare(Attendance o1, Attendance o2) {
-            return o1.getPeriod().getFrom() - o2.getPeriod().getTo();
+            return o1.getPeriod().getFrom().intValue()
+                    - o2.getPeriod().getTo().intValue();
         }
     }
